+++ tests/test_get_keys.py
@@ -5,6 +5,7 @@ import unittest
 import helpers
 
 import pycoreconf
+from pycoreconf.datastore import CORECONFDatastore
 
 
 class _DummyEnumKeyModel:
@@ -30,7 +31,7 @@ class _DummyEnumKeyModel:
 
 class TestGetKeys(unittest.TestCase):
     def test_enum_key_is_converted_both_directions(self):
-        ds = pycoreconf.CORECONFDatastore(_DummyEnumKeyModel(), {})
+        ds = CORECONFDatastore(_DummyEnumKeyModel(), {})
 
         sid, keys = ds._resolve_path("/root/list[mode='delta'][id='7']")
         self.assertEqual(sid, 110)